move utils to billingUtils and fix bug
