+++ src/lib/subscriberUtils.js
@@ -73,19 +73,4 @@ const getSubscriberTransactionData = async owner => {
   };
 };
 
-const processCurrencyInput = (event, balance, paymentAmount) => {
-  event.preventDefault();
-  let newAmount;
-  if (event.nativeEvent.key === 'Backspace') {
-    let paymentAmountInCents = paymentAmount * 100;
-    paymentAmountInCents = Math.floor(paymentAmountInCents / 10);
-    newAmount = paymentAmountInCents / 100;
-  } else if (/^[0123456789]$/.test(event.nativeEvent.key)) {
-    newAmount = paymentAmount * 10 + parseFloat(event.nativeEvent.key) / 100;
-  }
-  newAmount = Math.max(0.0, newAmount); // Pay atleast 1 cent
-  newAmount = Math.min(balance, newAmount); // Pay no more than balance
-  return newAmount.toFixed(2);
-};
-
-export { formatAmount, getSubscriberTransactionData, processCurrencyInput };
+export { formatAmount, getSubscriberTransactionData };
